refactor: Command enumを別モジュールに切り出す
